pyproject.toml
[tool.poetry]
name = "loom-bot"
version = "0.1.0"
description = "A smart LP."
authors = ["sunny <sunny@foil.xyz>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.15.1"
python-dotenv = "^1.0.1"
gql = {extras = ["requests"], version = "^3.5.0"}
requests = "^2.31.0"
typing-extensions = "^4.10.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.5.0"

[tool.poetry.scripts]
bot = "src.bot.__main__:main"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I"] 